move computed values into VitalsMetrics model
