+++ src/core/PageVitals/helpers.ts
@@ -14,67 +14,17 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
+import { DownloadableData, DownloadableDataset } from "./types";
 import {
   SummaryCard,
-  SummaryStatus,
-  VitalsSummaryTableRow,
   METRIC_TYPES,
   METRIC_TYPE_LABELS,
   MetricType,
-  DownloadableData,
-  DownloadableDataset,
-} from "./types";
-import {
   ENTITY_TYPES,
-  VitalsSummaryRecord,
   VitalsTimeSeriesRecord,
+  VitalsSummaryTableRow,
 } from "../models/types";
 import { formatPercent } from "../../utils/formatStrings";
-
-export function getSummaryStatus(value: number): SummaryStatus {
-  if (value < 70) return "POOR";
-  if (value >= 70 && value < 80) return "NEEDS_IMPROVEMENT";
-  if (value >= 80 && value < 90) return "GOOD";
-  if (value >= 90 && value < 95) return "GREAT";
-  return "EXCELLENT";
-}
-
-export const getSummaryCards: (
-  summary: VitalsSummaryRecord
-) => SummaryCard[] = (summary) => [
-  {
-    title: METRIC_TYPE_LABELS.OVERALL,
-    description: "Average timeliness across all metrics",
-    value: summary.overall,
-    status: getSummaryStatus(summary.overall),
-    id: METRIC_TYPES.OVERALL,
-  },
-  {
-    title: METRIC_TYPE_LABELS.DISCHARGE,
-    description: `of clients were discharged at their earliest projected regular
-     supervision discharge date`,
-    value: summary.timelyDischarge,
-    status: getSummaryStatus(summary.timelyDischarge),
-    id: METRIC_TYPES.DISCHARGE,
-  },
-  {
-    title: METRIC_TYPE_LABELS.CONTACT,
-    description: `of clients received initial contact within 30 days of starting
-     supervision and a F2F contact every subsequent 90, 60, or 30 days for 
-     minimum, medium, and maximum supervision levels respectively`,
-    value: summary.timelyContact,
-    status: getSummaryStatus(summary.timelyContact),
-    id: METRIC_TYPES.CONTACT,
-  },
-  {
-    title: METRIC_TYPE_LABELS.RISK_ASSESSMENT,
-    description: `of clients have had an initial assessment within 30 days and 
-      reassessment within 212 days`,
-    value: summary.timelyRiskAssessment,
-    status: getSummaryStatus(summary.timelyRiskAssessment),
-    id: METRIC_TYPES.RISK_ASSESSMENT,
-  },
-];
 
 export function getSummaryDetail(
   summaryCards: SummaryCard[],
@@ -83,39 +33,6 @@ export function getSummaryDetail(
   return (
     summaryCards.find((card) => card.id === selectedCardId) || summaryCards[0]
   );
-}
-
-export function getEntitySummaries(
-  vitalsSummaries: VitalsSummaryRecord[],
-  currentEntityId: string
-): {
-  currentEntitySummary: VitalsSummaryRecord;
-  childEntitySummaryRows: VitalsSummaryTableRow[];
-  parentEntityName?: string;
-} {
-  const currentEntitySummary = vitalsSummaries.find(
-    (d) => d.entityId === currentEntityId
-  ) as VitalsSummaryRecord;
-  const childEntitySummaryRows = vitalsSummaries
-    .filter(
-      (d) =>
-        d.parentEntityId === currentEntityId && d.parentEntityId !== d.entityId
-    )
-    .map((d) => {
-      const { entityId, entityName, entityType, ...attrs } = d;
-      return {
-        entity: {
-          entityId,
-          entityName,
-          entityType,
-        },
-        ...attrs,
-      };
-    }) as VitalsSummaryTableRow[];
-  const parentEntityName = vitalsSummaries.find(
-    (d) => d.entityId === currentEntitySummary.parentEntityId
-  )?.entityName;
-  return { currentEntitySummary, childEntitySummaryRows, parentEntityName };
 }
 
 export function getTimeSeries(
@@ -216,27 +133,3 @@ export function getVitalsSummaryDownloadableData(
     dataExportLabel,
   };
 }
-
-export const getVitalsFiltersText = (
-  currentEntitySummary: VitalsSummaryRecord,
-  children: VitalsSummaryTableRow[],
-  parentEntityName: string | undefined
-): string => {
-  let offices;
-  let officers;
-  switch (currentEntitySummary.entityType) {
-    case ENTITY_TYPES.LEVEL_1_SUPERVISION_LOCATION:
-      offices = currentEntitySummary.entityName;
-      officers = children.map((child) => child.entity.entityName).join(", ");
-      break;
-    case ENTITY_TYPES.PO:
-      offices = parentEntityName;
-      officers = currentEntitySummary.entityName;
-      break;
-    default:
-      offices = "All";
-      officers = "All";
-  }
-
-  return `Office(s): ${offices}, Officers: ${officers}`;
-};
