+++ src/core/models/__tests__/VitalsMetrics.test.ts
@@ -18,7 +18,7 @@
 // import CoreStore from "../../CoreStore";
 import { callMetricsApi } from "../../../api/metrics/metricsClient";
 import RootStore from "../../../RootStore";
-import VitalsMetrics from "../VitalsMetrics";
+import VitalsMetrics, { getSummaryStatus } from "../VitalsMetrics";
 
 const mockTenantId = "US_ND";
 
@@ -112,5 +112,43 @@ describe("VitalsMetrics", () => {
         monthlyAvg: 90.625,
       },
     ]);
+  });
+});
+
+describe("getSummaryStatus", () => {
+  describe("when value is less than 70", () => {
+    it("returns POOR", () => {
+      [0, 15, 25, 69].forEach((number) => {
+        expect(getSummaryStatus(number)).toEqual("POOR");
+      });
+    });
+  });
+  describe("when value is greater than or equal to 70 and less than 80", () => {
+    it("returns NEEDS_IMPROVEMENT", () => {
+      [70, 75, 79].forEach((number) => {
+        expect(getSummaryStatus(number)).toEqual("NEEDS_IMPROVEMENT");
+      });
+    });
+  });
+  describe("when value is greater than or equal to 80 and less than 90", () => {
+    it("returns GOOD", () => {
+      [80, 85, 89].forEach((number) => {
+        expect(getSummaryStatus(number)).toEqual("GOOD");
+      });
+    });
+  });
+  describe("when value is greater than or equal to 90 and less than 95", () => {
+    it("returns GREAT", () => {
+      [90, 94].forEach((number) => {
+        expect(getSummaryStatus(number)).toEqual("GREAT");
+      });
+    });
+  });
+  describe("when value is greater than 95", () => {
+    it("returns EXCELLENT", () => {
+      [95, 100].forEach((number) => {
+        expect(getSummaryStatus(number)).toEqual("EXCELLENT");
+      });
+    });
   });
 });
